requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
pymongo==4.3.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.24.2
joblib==1.2.0
flask-mail==0.9.1
gunicorn==20.1.0
bcrypt==4.0.1
python-jose==3.3.0
Werkzeug==2.2.3